requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sentry-sdk==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
